requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
falcon==1.4.1
Flask==1.0.2
Flask-Login==0.4.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
python-mimeparse==1.6.0
requests==2.21.0
secure-smtplib==0.1.1
six==1.12.0
SQLAlchemy==1.3.1
urllib3==1.24.1
Werkzeug==0.15.1

bs4<0.1.0,>=0.0.1
editdistance<0.6.0,>=0.5.2
lxml<5.0.0,>=4.2.5
nltk<4.0,>=3.3
numpy<2.0,>=1.11
pandas<0.24.0,>=0.23.4
pyyaml<5.0,>=4.2b1
scipy<2.0.0,>=1.1.0
snorkel-metal<0.5.0,>=0.4.1
spacy<2.2.0,>=2.1.3
sqlalchemy[postgresql_psycopg2binary]<2.0.0,>=1.2.12
tensorboardX<2.0,>=1.6
torch<2.0,>=1.0
tqdm<5.0.0,>=4.26.0
treedlib<0.2.0,>=0.1.1
wand<0.5.0,>=0.4.4
ipython
psqycopg2
flair
gunicorn